only delete checkpoint file if it's found

+++ moseq2_model/cli.py
@@ -187,7 +187,8 @@ def learn_model(input_file, dest_file, hold_out, hold_out_seed, nfolds, ncpus,
                 shutil.move(checkpoint_file, checkpoint_file + '.1')
             except FileNotFoundError:
                 print('Loading original checkpoint failed, checking backup')
-                os.remove(checkpoint_file)
+                if os.path.exists(checkpoint_file):
+                    os.remove(checkpoint_file)
                 checkpoint = load_arhmm_checkpoint(checkpoint_file2, train_data)
             arhmm = checkpoint.pop('model')
             print('On iteration', checkpoint['iter'])
